Implementing setTraits, which allows multiple webview settings to be changed at once.  These settings will be passed in via a JSON object. If a setting is missing in the object, it is unchanged in the webview. If a setting in the object is not yet implemented, it is ignored.
Also, there were difficulties working with the customHTTPHeader functions,
since they accept QString's but immediately convert them to char **.  It
seems that processing the strings in the char ** format is trickier than
it needs to be.  It makes more sense to process and store the strings as
QString's, and convert them when we actually need them to be char *.

Some of the unit tests also had to be fixed to work on my computer.